+++ lib/phone_numbers_parser.dart
@@ -3,4 +3,13 @@
 /// More dartdocs go here.
 library phone_number_parser;
 
-export 'src/phone_numbers_parser_base.dart';
+export 'src/parsers/light_phone_parser.dart';
+export 'src/parsers/phone_parser.dart';
+export 'src/parsers/base_phone_parser.dart';
+
+export 'src/formatters/phone_number_formatter.dart';
+
+export 'src/models/phone_number_type.dart';
+export 'src/models/phone_number.dart';
+export 'src/models/phone_number_exceptions.dart';
+export 'src/models/phone_number_range.dart';