+++ lib/src/parsers/light_phone_parser.dart
@@ -1,6 +1,6 @@
 import 'package:phone_numbers_parser/phone_numbers_parser.dart';
 import 'package:phone_numbers_parser/src/models/phone_number.dart';
-import 'package:phone_numbers_parser/src/parsers/_dial_code_parser.dart';
+import 'package:phone_numbers_parser/src/parsers/_country_code_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_international_prefix_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_iso_code_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_national_prefix_parser.dart';
@@ -31,26 +31,40 @@ class LightPhoneParser extends BasePhoneParser {
   ///
   /// throws a PhoneNumberException if the isoCode is invalid
   @override
-  PhoneNumber parseWithIsoCode(String isoCode, String phoneNumber) {
+  PhoneNumber parseWithIsoCode(
+    String isoCode,
+    String phoneNumber,
+  ) {
     isoCode = IsoCodeParser.normalizeIsoCode(isoCode);
     phoneNumber = TextParser.normalize(phoneNumber);
     final metadata = MetadataFinder.getMetadataForIsoCode(isoCode);
     phoneNumber =
         InternationalPrefixParser.removeInternationalPrefix(phoneNumber);
-    phoneNumber = DialCodeParser.removeDialCode(phoneNumber, metadata.dialCode);
+    phoneNumber =
+        CountryCodeParser.removeCountryCode(phoneNumber, metadata.countryCode);
     final nsn =
         NationalPrefixParser.removeNationalPrefix(phoneNumber, metadata);
     return PhoneNumber(nsn: nsn, isoCode: metadata.isoCode);
   }
 
   @override
-  PhoneNumber parseWithDialCode(String dialCode, String phoneNumber) {
-    dialCode = DialCodeParser.normalizeDialCode(dialCode);
+  @Deprecated(
+      'renamed to parse with countryCode, dial code was semantically incorrect')
+  PhoneNumber parseWithDialCodeCode(String dialCode, String phoneNumber) {
+    return parseWithCountryCode(dialCode, phoneNumber);
+  }
+
+  @override
+  PhoneNumber parseWithCountryCode(
+    String countryCode,
+    String phoneNumber,
+  ) {
+    countryCode = CountryCodeParser.normalizeCountryCode(countryCode);
     phoneNumber = TextParser.normalize(phoneNumber);
     phoneNumber =
         InternationalPrefixParser.removeInternationalPrefix(phoneNumber);
-    phoneNumber = DialCodeParser.removeDialCode(phoneNumber, dialCode);
-    var metadatas = MetadataFinder.getMetadatasForDialCode(dialCode);
+    phoneNumber = CountryCodeParser.removeCountryCode(phoneNumber, countryCode);
+    var metadatas = MetadataFinder.getMetadatasForCountryCode(countryCode);
     metadatas =
         MetadataMatcher.reducePotentialMetadatasFits(phoneNumber, metadatas);
     // this line will give some fake results between US and CA but that's the
@@ -66,8 +80,8 @@ class LightPhoneParser extends BasePhoneParser {
     phoneNumber = TextParser.normalize(phoneNumber);
     phoneNumber =
         InternationalPrefixParser.removeInternationalPrefix(phoneNumber);
-    final dialCode = DialCodeParser.extractDialCode(phoneNumber);
-    return parseWithDialCode(dialCode, phoneNumber);
+    final countryCode = CountryCodeParser.extractCountryCode(phoneNumber);
+    return parseWithCountryCode(countryCode, phoneNumber);
   }
 
   /// Validates a phone number using length information